+++ VersionHistory.md
@@ -7,11 +7,13 @@ is updated in preparation for publishing an updated NuGet package.
 
 Prefix the description of the change with `[major]`, `[minor]` or `[patch]` in accordance with [SemVer](http://semver.org).
 
+## Released
+
+### 0.5.0
+
 * [major] Remove `EnumerableUtility.ToHashSet()`.
 * [major] Seal `WrappingStream`.
 * [major] Introduce `WrappingStreamBase` and make it the base class of `CachingStream`, `ReadOnlyStream`, `RebasedStream`, and `TruncatedStream`. This fixes `CopyToAsync` for `RebasedStream` and `TruncatedStream`.
-
-## Released
 
 ### 0.4.2